Move some functions from `replay.utils` to `replay.ann.index_stores`
