+++ replay/ann/index_stores/hdfs_index_store.py
@@ -7,7 +7,7 @@ from typing import Callable, Any
 from pyarrow import fs
 
 from replay.ann.index_stores.base_index_store import IndexStore
-from replay.utils import get_filesystem, FileSystem
+from replay.ann.index_stores.utils import FileSystem, get_filesystem
 
 logger = logging.getLogger("replay")
